+++ src/lib.rs
@@ -24,8 +24,8 @@
 //!
 //! let graph = Graph {
 //!     nodes: vec![
-//!         Node { id: NodeId(0), radius: 1 },
-//!         Node { id: NodeId(1), radius: 1 },
+//!         Node { id: NodeId(0), width: 3, height: 3, fixed: None },
+//!         Node { id: NodeId(1), width: 3, height: 3, fixed: None },
 //!     ],
 //!     edges: vec![Edge {
 //!         id: EdgeId(0),